- Javadoc User and Alarm - User stored in Firebase under DATA tag (change User.toMap()) - Sport variables change to Objects (from int, float.. to Integer, Double..) - Now Event is not Parcelable (there is no reason to be) - Comment about why Field, Sport, SportCourt and MyPlace are Parcelable

+++ app/src/main/java/com/usal/jorgeav/sportapp/adapters/AddSportsAdapter.java
@@ -78,7 +78,7 @@ public class AddSportsAdapter extends RecyclerView.Adapter<AddSportsAdapter.View
             int nameResource = MyApplication.getAppContext().getResources()
                     .getIdentifier(s.getSportID() , "string", MyApplication.getAppContext().getPackageName());
             holder.textViewSportName.setText(nameResource);
-            holder.ratingBarSportLevel.setRating(s.getPunctuation());
+            holder.ratingBarSportLevel.setRating(s.getPunctuation().floatValue());
         }
     }
 
@@ -143,7 +143,7 @@ public class AddSportsAdapter extends RecyclerView.Adapter<AddSportsAdapter.View
             ratingBarSportLevel.setOnRatingBarChangeListener(new RatingBar.OnRatingBarChangeListener() {
                 @Override
                 public void onRatingChanged(RatingBar ratingBar, float v, boolean b) {
-                    mDataset.get(getAdapterPosition()).setPunctuation(v);
+                    mDataset.get(getAdapterPosition()).setPunctuation((double) v);
                 }
             });
         }
